Build a shim binary that converts node-config.yaml to kubelet args
Will remove the need to have the kubelet start from openshift master
