+++ pkg/cmd/server/start/start_node.go
@@ -7,8 +7,6 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"regexp"
-	"strconv"
 	"strings"
 	"syscall"
 
@@ -35,6 +33,7 @@ import (
 	networkoptions "github.com/openshift/origin/pkg/cmd/server/kubernetes/network/options"
 	"github.com/openshift/origin/pkg/cmd/server/kubernetes/node"
 	nodeoptions "github.com/openshift/origin/pkg/cmd/server/kubernetes/node/options"
+	originnode "github.com/openshift/origin/pkg/cmd/server/origin/node"
 	cmdutil "github.com/openshift/origin/pkg/cmd/util"
 	utilflags "github.com/openshift/origin/pkg/cmd/util/flags"
 	"github.com/openshift/origin/pkg/version"
@@ -281,7 +280,7 @@ func (o NodeOptions) RunNode() error {
 	}
 
 	if o.NodeArgs.WriteFlagsOnly {
-		return WriteKubeletFlags(*nodeConfig)
+		return originnode.WriteKubeletFlags(*nodeConfig)
 	}
 
 	return StartNode(*nodeConfig, o.NodeArgs.Components)
@@ -425,38 +424,6 @@ func execKubelet(kubeletArgs []string) error {
 	// execve the child process, replacing this process
 	glog.V(3).Infof("Exec %s %s", kubeletPath, strings.Join(args, " "))
 	return syscall.Exec(kubeletPath, args, os.Environ())
-}
-
-// safeArgRegexp matches only characters that are known safe. DO NOT add to this list
-// without fully considering whether that new character can be used to break shell escaping
-// rules.
-var safeArgRegexp = regexp.MustCompile(`^[\da-zA-Z\-=_\.,/\:]+$`)
-
-// shellEscapeArg quotes an argument if it contains characters that my cause a shell
-// interpreter to split the single argument into multiple.
-func shellEscapeArg(s string) string {
-	if safeArgRegexp.MatchString(s) {
-		return s
-	}
-	return strconv.Quote(s)
-}
-
-// WriteKubeletFlags writes the correct set of flags to start a Kubelet from the provided node config to
-// stdout, instead of launching anything.
-func WriteKubeletFlags(nodeConfig configapi.NodeConfig) error {
-	kubeletArgs, err := nodeoptions.ComputeKubeletFlags(nodeConfig.KubeletArguments, nodeConfig)
-	if err != nil {
-		return fmt.Errorf("cannot create kubelet args: %v", err)
-	}
-	if err := nodeoptions.CheckFlags(kubeletArgs); err != nil {
-		return err
-	}
-	var outputArgs []string
-	for _, s := range kubeletArgs {
-		outputArgs = append(outputArgs, shellEscapeArg(s))
-	}
-	fmt.Println(strings.Join(outputArgs, " "))
-	return nil
 }
 
 // StartNode launches the node processes.
